requirements.txt
python-dotenv==0.20.0
toml==0.10.2
python-telegram-bot==13.11
Flask==2.1.1
RPi.GPIO==0.7.1